Move diagnostic processing into a library
The diagnostics are going to be used throughout the whole system, so
there is little reason to keep them in the parser crate. Move them into
their own library. It's named 'locus' because it deals with locations.
