+++ src/README.md
@@ -3,5 +3,6 @@ Sabre source code
 
 Essential source code of Sabre.
 
+- `liblocus` — handling code locations and diagnostics
 - `libreader` — S-expression parser
 - `libunicode` — Unicode support
