+++ src/liblocus/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "liblocus"
+version = "0.1.0"
+authors = ["Sabre developers"]
+license = "MIT/Apache-2.0"
+
+[lib]
+name = "locus"
+path = "lib.rs"
