+++ src/liblocus/README.md
@@ -1,0 +1,4 @@
+liblocus
+========
+
+Here we have everything related to source code locations: spans, diagnostics.
