+++ src/liblocus/lib.rs
@@ -1,0 +1,14 @@
+// Copyright (c) 2017, Sabre developers
+//
+// Licensed under the Apache License, Version 2.0 (see LICENSE.Apache in the
+// root directory) or MIT license (see LICENSE.MIT in the root directory),
+// at your option. This file may be copied, distributed, and modified only
+// in accordance with the terms specified by the chosen license.
+
+//! Source code locations.
+//!
+//! This crate deals with source code location information like code spans,
+//! diagnostics.
+
+pub mod diagnostics;
+mod diagnostic_kinds;
