+++ src/libreader/lib.rs
@@ -9,11 +9,12 @@
 //!
 //! This crate contains modules implementing basic syntax analysis of Scheme source code.
 
+extern crate locus;
+
 #[cfg(feature = "unicode")]
 extern crate unicode;
 
 pub mod datum;
-pub mod diagnostics;
 pub mod intern_pool;
 pub mod lexer;
 pub mod parser;
